cloud_function/open_weather/requirements.txt
requests
datetime
google-cloud
google-cloud-bigquery
google-cloud-storage